Module aws_sdk_efs::error

Modules

access_point_already_exists

See AccessPointAlreadyExists

access_point_limit_exceeded

See AccessPointLimitExceeded

access_point_not_found

See AccessPointNotFound

availability_zones_mismatch

See AvailabilityZonesMismatch

bad_request

See BadRequest

dependency_timeout

See DependencyTimeout

file_system_already_exists

See FileSystemAlreadyExists

file_system_in_use

See FileSystemInUse

file_system_limit_exceeded

See FileSystemLimitExceeded

file_system_not_found

See FileSystemNotFound

incorrect_file_system_life_cycle_state

See IncorrectFileSystemLifeCycleState

incorrect_mount_target_state

See IncorrectMountTargetState

insufficient_throughput_capacity

See InsufficientThroughputCapacity

internal_server_error

See InternalServerError

invalid_policy_exception

See InvalidPolicyException

ip_address_in_use

See IpAddressInUse

mount_target_conflict

See MountTargetConflict

mount_target_not_found

See MountTargetNotFound

network_interface_limit_exceeded

See NetworkInterfaceLimitExceeded

no_free_addresses_in_subnet

See NoFreeAddressesInSubnet

policy_not_found

See PolicyNotFound

security_group_limit_exceeded

See SecurityGroupLimitExceeded

security_group_not_found

See SecurityGroupNotFound

subnet_not_found

See SubnetNotFound

throughput_limit_exceeded

See ThroughputLimitExceeded

too_many_requests

See TooManyRequests

unsupported_availability_zone

See UnsupportedAvailabilityZone

validation_exception

See ValidationException

Structs

AccessPointAlreadyExists

Returned if the access point you are trying to create already exists, with the

AccessPointLimitExceeded

Returned if the Amazon Web Services account has already created the maximum number of access points

AccessPointNotFound

Returned if the specified AccessPointId value doesn't exist in the

AvailabilityZonesMismatch

Returned if the Availability Zone that was specified for a mount target is different from the Availability Zone that was specified for One Zone storage classes.

BadRequest

Returned if the request is malformed or contains an error such as an invalid

CreateAccessPointError

CreateFileSystemError

CreateMountTargetError

CreateTagsError

DeleteAccessPointError

DeleteFileSystemError

DeleteFileSystemPolicyError

DeleteMountTargetError

DeleteTagsError

DependencyTimeout

The service timed out trying to fulfill the request, and the client should try the

DescribeAccessPointsError

DescribeAccountPreferencesError

DescribeBackupPolicyError

DescribeFileSystemPolicyError

DescribeFileSystemsError

DescribeLifecycleConfigurationError

DescribeMountTargetSecurityGroupsError

DescribeMountTargetsError

DescribeTagsError

FileSystemAlreadyExists

Returned if the file system you are trying to create already exists, with the

FileSystemInUse

Returned if a file system has mount targets.

FileSystemLimitExceeded

Returned if the Amazon Web Services account has already created the maximum number of file systems

FileSystemNotFound

Returned if the specified FileSystemId value doesn't exist in the

IncorrectFileSystemLifeCycleState

Returned if the file system's lifecycle state is not "available".

IncorrectMountTargetState

Returned if the mount target is not in the correct state for the

InsufficientThroughputCapacity

Returned if there's not enough capacity to provision additional throughput. This value

InternalServerError

Returned if an error occurred on the server side.

InvalidPolicyException

Returned if the FileSystemPolicy is is malformed or contains an error such as an invalid

IpAddressInUse

Returned if the request specified an IpAddress that is already in use

ListTagsForResourceError

ModifyMountTargetSecurityGroupsError

MountTargetConflict

Returned if the mount target would violate one of the specified restrictions based

MountTargetNotFound

Returned if there is no mount target with the specified ID found in the

NetworkInterfaceLimitExceeded

The calling account has reached the limit for elastic network interfaces for the

NoFreeAddressesInSubnet

Returned if IpAddress was not specified in the request and there are

PolicyNotFound

Returned if the default file system policy is in effect for the EFS file system specified.

PutAccountPreferencesError

PutBackupPolicyError

PutFileSystemPolicyError

PutLifecycleConfigurationError

SecurityGroupLimitExceeded

Returned if the size of SecurityGroups specified in the request is

SecurityGroupNotFound

Returned if one of the specified security groups doesn't exist in the subnet's

SubnetNotFound

Returned if there is no subnet with ID SubnetId provided in the

TagResourceError

ThroughputLimitExceeded

Returned if the throughput mode or amount of provisioned throughput can't be changed

TooManyRequests

Returned if you don’t wait at least 24 hours before changing the throughput mode, or

UnsupportedAvailabilityZone

Returned if the requested Amazon EFS functionality is not available in the specified Availability Zone.

UntagResourceError

UpdateFileSystemError

ValidationException

Returned if the Backup service is not available in the Amazon Web Services Region in which the request was made.

Enums

CreateAccessPointErrorKind

CreateFileSystemErrorKind

CreateMountTargetErrorKind

CreateTagsErrorKind

DeleteAccessPointErrorKind

DeleteFileSystemErrorKind

DeleteFileSystemPolicyErrorKind

DeleteMountTargetErrorKind

DeleteTagsErrorKind

DescribeAccessPointsErrorKind

DescribeAccountPreferencesErrorKind

DescribeBackupPolicyErrorKind

DescribeFileSystemPolicyErrorKind

DescribeFileSystemsErrorKind

DescribeLifecycleConfigurationErrorKind

DescribeMountTargetSecurityGroupsErrorKind

DescribeMountTargetsErrorKind

DescribeTagsErrorKind

ListTagsForResourceErrorKind

ModifyMountTargetSecurityGroupsErrorKind

PutAccountPreferencesErrorKind

PutBackupPolicyErrorKind

PutFileSystemPolicyErrorKind

PutLifecycleConfigurationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateFileSystemErrorKind