Module aws_sdk_efs::model

Modules

access_point_description

See AccessPointDescription

backup_policy

See BackupPolicy

creation_info

See CreationInfo

file_system_description

See FileSystemDescription

file_system_size

See FileSystemSize

lifecycle_policy

See LifecyclePolicy

mount_target_description

See MountTargetDescription

posix_user

See PosixUser

resource_id_preference

See ResourceIdPreference

root_directory

See RootDirectory

tag

See Tag

Structs

AccessPointDescription

Provides a description of an EFS file system access point.

BackupPolicy

The backup policy for the file system used to create automatic daily backups. If status has a value of

CreationInfo

Required if the RootDirectory > Path specified does not exist.

FileSystemDescription

A description of the file system.

FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its

LifecyclePolicy

Describes a policy used by EFS lifecycle management to transition files to the Infrequent

MountTargetDescription

Provides a description of a mount target.

PosixUser

The full POSIX identity, including the user ID, group ID, and any secondary group IDs, on the access point that is used for all file system operations performed by

ResourceIdPreference

Describes the resource type and its ID preference for the user's Amazon Web Services account, in the current Amazon Web Services Region.

RootDirectory

Specifies the directory on the Amazon EFS file system that the access point provides access to.

Tag

A tag is a key-value pair. Allowed characters are letters, white space, and numbers that

Enums

LifeCycleState

PerformanceMode

Resource

An EFS resource, for example a file system or a mount target.

ResourceIdType

A preference indicating a choice to use 63bit/32bit IDs for all applicable resources.

Status

ThroughputMode

TransitionToIaRules

TransitionToPrimaryStorageClassRules