Module aws_sdk_efs::operation

Structs

CreateAccessPoint

Creates an EFS access point. An access point is an application-specific view into an EFS

CreateFileSystem

Creates a new, empty file system. The operation requires a creation token in the

CreateMountTarget

Creates a mount target for a file system. You can then mount the file system on EC2

CreateTags

DeleteAccessPoint

Deletes the specified access point. After deletion is complete, new clients can no

DeleteFileSystem

Deletes a file system, permanently severing access to its contents. Upon return, the

DeleteFileSystemPolicy

Deletes the FileSystemPolicy for the specified file system.

DeleteMountTarget

Deletes the specified mount target.

DeleteTags

DescribeAccessPoints

Returns the description of a specific Amazon EFS access point if the AccessPointId is provided.

DescribeAccountPreferences

Returns the account preferences settings for the Amazon Web Services account associated with the user making the request, in the current Amazon Web Services Region.

DescribeBackupPolicy

Returns the backup policy for the specified EFS file system.

DescribeFileSystemPolicy

Returns the FileSystemPolicy for the specified EFS file system.

DescribeFileSystems

Returns the description of a specific Amazon EFS file system if either the file system

DescribeLifecycleConfiguration

Returns the current LifecycleConfiguration object for the specified Amazon

DescribeMountTargetSecurityGroups

Returns the security groups currently in effect for a mount target. This operation

DescribeMountTargets

Returns the descriptions of all the current mount targets, or a specific mount target,

DescribeTags

ListTagsForResource

Lists all tags for a top-level EFS resource. You must provide the ID of the resource that you want to retrieve the tags for.

ModifyMountTargetSecurityGroups

Modifies the set of security groups in effect for a mount target.

PutAccountPreferences

Use this operation to set the account preference in the current Amazon Web Services Region to use either long 17 character (63 bit) or short 8 character (32 bit) IDs for

PutBackupPolicy

Updates the file system's backup policy. Use this action to start or stop automatic backups of the file system.

PutFileSystemPolicy

Applies an Amazon EFS FileSystemPolicy to an Amazon EFS file system.

PutLifecycleConfiguration

Enables lifecycle management by creating a new LifecycleConfiguration

TagResource

Creates a tag for an EFS resource. You can create tags for EFS file systems and access points using this API operation.

UntagResource

Removes tags from an EFS resource. You can remove tags from EFS file systems and access points using this API operation.

UpdateFileSystem

Updates the throughput mode or the amount of provisioned throughput of an existing file